fix to the crud authz

+++ dotnet/AuthzDeviceGateway/DocumentsWebApi/Controllers/DocumentsController.cs
@@ -7,6 +7,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using DocumentsWebApi.Data;
 using DocumentsWebApi.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using DocumentsWebApi.Authorization;
+using System.Security.Claims;
 
 namespace DocumentsWebApi.Controllers
 {
@@ -17,22 +21,35 @@ namespace DocumentsWebApi.Controllers
         private const string FilesFolder = "_files";
         private readonly ILogger<DocumentsController> _logger;
         private readonly DocumentsDbContext _context;
+        private readonly IAuthorizationService _authorizationService;
         private string _fullPath;
 
         public DocumentsController(
             ILogger<DocumentsController> logger,
-            DocumentsDbContext context)
+            DocumentsDbContext context,
+            IAuthorizationService authorizationService)
         {
             _logger = logger;
             _context = context;
-
+            _authorizationService = authorizationService;
             _fullPath = Path.GetFullPath(FilesFolder);
         }
 
         // GET: api/Documents
         [HttpGet]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme
+            //, Policy = Policies.DocsList  // not-templated => always fails
+            )]
         public async Task<ActionResult<IEnumerable<FullDocument>>> GetDocuments()
         {
+            var authResult = await _authorizationService.AuthorizeAsync(User,
+                //Policies.DocsList); // not-templated => always fails
+                Document.Empty, Operations.List);
+            if (!authResult.Succeeded)
+            {
+                return Unauthorized();
+            }
+
             return await _context.Documents
                 .Select(d => new FullDocument(d, string.Empty))
                 .ToListAsync();
@@ -40,6 +57,7 @@ namespace DocumentsWebApi.Controllers
 
         // GET: api/Documents/5
         [HttpGet("{id}")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<ActionResult<FullDocument>> GetDocument(Guid id)
         {
             var document = await _context.Documents.FindAsync(id);
@@ -47,6 +65,13 @@ namespace DocumentsWebApi.Controllers
             if (document == null)
             {
                 return NotFound();
+            }
+
+            var authResult = await _authorizationService.AuthorizeAsync(
+                User, document, Operations.Read);
+            if (!authResult.Succeeded)
+            {
+                return Unauthorized();
             }
 
             //document = await PatchFilename(document);
@@ -59,6 +84,7 @@ namespace DocumentsWebApi.Controllers
         // PUT: api/Documents/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<IActionResult> PutDocument(Guid id, FullDocument fullDocument)
         {
             if (id != fullDocument.Document.Id)
@@ -66,7 +92,25 @@ namespace DocumentsWebApi.Controllers
                 return BadRequest();
             }
 
-            _context.Entry(fullDocument.Document).State = EntityState.Modified;
+            var current = await _context.Documents.FindAsync(id);
+            if (current == null)
+            {
+                return NotFound();
+            }
+
+            // authorization is applied on the document info saved on the DB
+            // the document coming from the user may have been tampered
+            var authResult = await _authorizationService.AuthorizeAsync(
+                User, current, Operations.Update);
+            if (!authResult.Succeeded)
+            {
+                return Unauthorized();
+            }
+
+            // we only copy the modifiable data in the current document
+            current.Name = fullDocument.Document.Name;
+            current.Description = fullDocument.Document.Name;
+            _context.Entry(current).State = EntityState.Modified;
 
             try
             {
@@ -92,13 +136,22 @@ namespace DocumentsWebApi.Controllers
         // POST: api/Documents
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme,
+            Policy = Policies.DocsCreate)]
         public async Task<ActionResult<FullDocument>> PostDocument(FullDocument fullDocument)
         {
             _context.Documents.Add(fullDocument.Document);
             var filename = $"Doc_{fullDocument.Document.Id.ToString().ToLower()}";
             fullDocument.Document.Pathname = filename;
-            await _context.SaveChangesAsync();
 
+            var email = ((ClaimsIdentity?)User.Identity)
+                ?.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Email)
+                ?.Value;
+            if (email == null) return Unauthorized();
+
+            fullDocument.Document.Author = email;
+
+            await _context.SaveChangesAsync();
             await SaveMarkdown(filename, fullDocument.Markdown);
 
             return CreatedAtAction("GetDocument",
@@ -107,12 +160,20 @@ namespace DocumentsWebApi.Controllers
 
         // DELETE: api/Documents/5
         [HttpDelete("{id}")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<IActionResult> DeleteDocument(Guid id)
         {
             var document = await _context.Documents.FindAsync(id);
             if (document == null)
             {
                 return NotFound();
+            }
+
+            var authResult = await _authorizationService.AuthorizeAsync(
+                User, document, Operations.Delete);
+            if (!authResult.Succeeded)
+            {
+                return Unauthorized();
             }
 
             _context.Documents.Remove(document);
